move Default Template to Feature level settings
