+++ includes/Classifai/Features/RecommendedContent.php
@@ -42,6 +42,8 @@ class RecommendedContent extends Feature {
 	public function feature_setup() {
 		$settings = $this->get_settings();
 
+		add_action( 'admin_enqueue_scripts', [ $this, 'enqueue_feature_setting_assets' ] );
+
 		if ( isset( $settings['provider'] ) && OpenAIEmbeddings::ID === $settings['provider'] ) {
 			add_action( 'enqueue_block_editor_assets', [ $this, 'enqueue_editor_assets' ] );
 			add_filter( 'pre_render_block', [ $this, 'pre_render_block' ], 10, 2 );
@@ -68,6 +70,19 @@ class RecommendedContent extends Feature {
 			'provider'         => OpenAIEmbeddings::ID,
 			'default_template' => 'title-date',
 		];
+	}
+
+	/**
+	 * Enqueues feature-level assets.
+	 */
+	public function enqueue_feature_setting_assets() {
+		wp_enqueue_script(
+			'classifai-plugin-recommended-content-feature-fields',
+			CLASSIFAI_PLUGIN_URL . 'dist/classifai-plugin-recommended-content-feature-fields.js',
+			get_asset_info( 'classifai-plugin-recommended-content-feature-fields', 'dependencies' ),
+			get_asset_info( 'classifai-plugin-recommended-content-feature-fields', 'version' ),
+			true
+		);
 	}
 
 	/**
