+++ webpack.config.js
@@ -43,6 +43,7 @@ module.exports = {
 			'./src/js/features/image-generation/extend-image-block-generate-image.js'
 		],
 		'classifai-plugin-image-generation-generate-image-media-upload': './src/js/features/image-generation/media-modal/views/generate-image-media-upload.js',
+		'classifai-plugin-recommended-content-feature-fields': './src/js/features/recommended-content/feature-fields-plugin.js',
 		settings: './src/js/settings/index.js',
 	},
 	module: {
